separate spindle controller class into separate file
